Restructure job dir to accommodate SS run
