+++ job/steps/collaborative_filtering/fetch_data.py
@@ -12,7 +12,7 @@ from typing import List, Set
 import pandas as pd
 
 from job.helpers import chunk_name
-from job.steps import warehouse
+from job.steps.collaborative_filtering import warehouse
 from lib.events import Event
 from lib.metrics import Unit, write_metric
 from sites.site import Site, get_bucket_name
